♻️ slash command builder
