+++ src/plugins/settings/index.ts
@@ -15,111 +15,9 @@ export default {
   data: new SlashCommandBuilder()
     .setName("settings")
     .setDescription("Manage settings.")
-    .addSubcommandGroup((group) =>
-      group
-        .setName("guild")
-        .setDescription("Manage guild settings.")
-        .addSubcommand((command) =>
-          command
-            .setName("pterodactyl")
-            .setDescription("Controlpanel.gg")
-            .addStringOption((option) =>
-              option
-                .setName("url")
-                .setDescription("The api url")
-                .setRequired(true)
-            )
-            .addStringOption((option) =>
-              option
-                .setName("token")
-                .setDescription("The api token")
-                .setRequired(true)
-            )
-        )
-        .addSubcommand((command) =>
-          command
-            .setName("credits")
-            .setDescription("Credits")
-            .addBooleanOption((option) =>
-              option
-                .setName("status")
-                .setDescription("Should credits be enabled?")
-            )
-            .addNumberOption((option) =>
-              option
-                .setName("rate")
-                .setDescription("Amount of credits per message.")
-            )
-            .addNumberOption((option) =>
-              option
-                .setName("minimum-length")
-                .setDescription("Minimum length of message to earn credits.")
-            )
-            .addNumberOption((option) =>
-              option
-                .setName("work-rate")
-                .setDescription("Maximum amount of credits on work.")
-            )
-            .addNumberOption((option) =>
-              option
-                .setName("work-timeout")
-                .setDescription(
-                  "Timeout between work schedules (milliseconds)."
-                )
-            )
-            .addNumberOption((option) =>
-              option
-                .setName("timeout")
-                .setDescription(
-                  "Timeout between earning credits (milliseconds)."
-                )
-            )
-        )
-        .addSubcommand((command) =>
-          command
-            .setName("points")
-            .setDescription("Points")
-            .addBooleanOption((option) =>
-              option
-                .setName("status")
-                .setDescription("Should credits be enabled?")
-            )
-            .addNumberOption((option) =>
-              option
-                .setName("rate")
-                .setDescription("Amount of credits per message.")
-            )
-            .addNumberOption((option) =>
-              option
-                .setName("minimum-length")
-                .setDescription("Minimum length of message to earn credits.")
-            )
-            .addNumberOption((option) =>
-              option
-                .setName("timeout")
-                .setDescription(
-                  "Timeout between earning credits (milliseconds)."
-                )
-            )
-        )
-    )
-    .addSubcommandGroup((group) =>
-      group
-        .setName("user")
-        .setDescription("Manage user settings.")
-        .addSubcommand((command) =>
-          command
-            .setName("appearance")
-            .setDescription("Manage your appearance")
-            .addStringOption((option) =>
-              option
-                .setName("language")
-                .setDescription("Configure your language")
-                .addChoice("English", "en")
-                .addChoice("Swedish", "sv")
-            )
-        )
-    ),
+    .addSubcommandGroup(guildGroup.data)
+    .addSubcommandGroup(userGroup.data),
+
   async execute(interaction: CommandInteraction) {
     // Destructure
     const { options, commandName, user, guild } = interaction;
@@ -127,12 +25,12 @@ export default {
     // Group - Guild
     if (options.getSubcommandGroup() === "guild") {
       // Execute Group - Guild
-      await guildGroup(interaction);
+      await guildGroup.execute(interaction);
     }
     // Group - User
     else if (options.getSubcommandGroup() === "user") {
       // Execute Group - User
-      await userGroup(interaction);
+      await userGroup.execute(interaction);
     }
 
     // Send debug message
